#1127: created all new transformers and query args, fixed old errors, but still have three errors left relating to the enums
