+++ src/backend/src/prisma-query-args/reimbursement-products.query-args.ts
@@ -1,0 +1,9 @@
+import { Prisma } from '@prisma/client';
+
+const reimbursementProductQueryArgs = Prisma.validator<Prisma.Reimbursement_ProductArgs>()({
+  include: {
+    wbsElement: true
+  }
+});
+
+export default reimbursementProductQueryArgs;
