+++ src/backend/src/transformers/expense-type.transformer.ts
@@ -1,0 +1,13 @@
+import { Prisma } from '@prisma/client';
+import { ExpenseType } from 'shared';
+
+const expenseTypeTransformer = (expenseType: Prisma.Expense_TypeGetPayload<null>): ExpenseType => {
+  return {
+    expenseTypeId: expenseType.expenseTypeId,
+    name: expenseType.name,
+    code: expenseType.code,
+    allowed: expenseType.allowed
+  };
+};
+
+export default expenseTypeTransformer;
