+++ src/backend/src/transformers/reimbursement-requests.transformer.ts
@@ -1,14 +1,15 @@
 import { Prisma } from '@prisma/client';
 import { ReimbursementRequest } from 'shared';
 import reimbursementRequestQueryArgs from '../prisma-query-args/reimbursement-requests.query-args';
-import { wbsNumOf } from '../utils/utils';
+import expenseTypeTransformer from './expense-type.transformer';
+import reimbursementProductTransformer from './reimbursement-products.transformer';
 import reimbursementStatusTransformer from './reimbursement-statuses.transformer';
 import userTransformer from './user.transformer';
+import vendorTransformer from './vendor.transformer';
 
 const reimbursementRequestTransformer = (
   reimbursementRequest: Prisma.Reimbursement_RequestGetPayload<typeof reimbursementRequestQueryArgs>
 ): ReimbursementRequest => {
-  const { vendor, expenseType } = reimbursementRequest;
   return {
     reimbursementRequestId: reimbursementRequest.reimbursementRequestId,
     saboId: reimbursementRequest.saboId ?? undefined,
@@ -17,28 +18,13 @@ const reimbursementRequestTransformer = (
     dateOfExpense: reimbursementRequest.dateOfExpense,
     reimbursementsStatuses: reimbursementRequest.reimbursementsStatuses.map(reimbursementStatusTransformer),
     recepient: userTransformer(reimbursementRequest.recepient),
-    vendor: {
-      vendorId: vendor.vendorId,
-      dateCreated: vendor.dateCreated,
-      name: vendor.name
-    },
+    vendor: vendorTransformer(reimbursementRequest.vendor),
     account: reimbursementRequest.account,
     totalCost: reimbursementRequest.totalCost,
     receiptPictures: reimbursementRequest.receiptPictures,
-    reimbursementProducts: reimbursementRequest.reimbursementProducts.map((reimbursementProduct) => ({
-      reimbursementProductId: reimbursementProduct.reimbursementProductId,
-      name: reimbursementProduct.name,
-      dateDeleted: reimbursementProduct.dateDeleted,
-      cost: reimbursementProduct.cost,
-      wbsNum: wbsNumOf(reimbursementProduct.wbsElement)
-    })),
+    reimbursementProducts: reimbursementRequest.reimbursementProducts.map(reimbursementProductTransformer),
     dateDelivered: reimbursementRequest.dateDelivered ?? undefined,
-    expenseType: {
-      expenseTypeId: expenseType.expenseTypeId,
-      name: expenseType.name,
-      code: expenseType.code,
-      allowed: expenseType.allowed
-    }
+    expenseType: expenseTypeTransformer(reimbursementRequest.expenseType)
   };
 };
